+++ app/src/main/java/com/chydee/notekeeper/NoteDao.java
@@ -7,6 +7,8 @@ import androidx.room.Insert;
 import androidx.room.Query;
 import androidx.room.Update;
 
+import com.chydee.notekeeper.model.Note;
+
 import java.util.List;
 
 @Dao
