+++ app/src/main/java/com/chydee/notekeeper/NoteViewModel.java
@@ -6,6 +6,7 @@ import androidx.annotation.NonNull;
 import androidx.lifecycle.AndroidViewModel;
 import androidx.lifecycle.LiveData;
 
+import com.chydee.notekeeper.model.Note;
 import com.chydee.notekeeper.repository.NoteRepository;
 
 import java.util.List;
