+++ app/src/main/java/com/chydee/notekeeper/database/NoteDatabase.java
@@ -9,8 +9,8 @@ import androidx.room.Room;
 import androidx.room.RoomDatabase;
 import androidx.sqlite.db.SupportSQLiteDatabase;
 
-import com.chydee.notekeeper.Note;
 import com.chydee.notekeeper.NoteDao;
+import com.chydee.notekeeper.model.Note;
 
 @Database(entities = {Note.class}, version = 1)
 public abstract class NoteDatabase extends RoomDatabase {
