+++ app/src/main/java/com/chydee/notekeeper/repository/NoteRepository.java
@@ -5,9 +5,9 @@ import android.os.AsyncTask;
 
 import androidx.lifecycle.LiveData;
 
-import com.chydee.notekeeper.Note;
 import com.chydee.notekeeper.NoteDao;
 import com.chydee.notekeeper.database.NoteDatabase;
+import com.chydee.notekeeper.model.Note;
 
 import java.util.List;
